Fix: private fields cannot be accessed by class methods.
isMethod is only true if the function is a method within the class' defintion.
If a function is 'assigned' to the instance as a method on it, its isMethod field will be false, even though it can be called on the instance like any other method.
Also made the printed text for methods different for variety and expressiveness.