Fix: functions without return values implicitly return a tuple.
Since we don't have any tuple-type objects in our language (internally), getting a tuple object as the return type indicates that the function is void (does not return anything).
This is better than the previous solution (checking if it returns nil), since nil may still be a valid value for a function to return, and should thus be treated like any other for call expressions.