Fix: scoping for local imports with shadowing.
When importing functions/objects/etc., we simply bind those objects to variables in the current scope. All scope and shadowing rules then apply to them as if they were regular variables.
This new approach avoids having the imported environments be added to the original varEnvs list, which originally prevented proper shadowing and complicated multiple parts in the Interpreter.