Implement private methods.
A method is private if the keyword 'safe' is placed before the function name. For practical reasons, constructors and static methods cannot be made private.
The LoxFunction constructor was also simplified to use a dictionary of relevant values, rather than passing each one to it independently.
The main change here was to split class methods into private and public dictionaries, and only allow access to methods in the private dictionary from within a method of the same class.