Replace increment operators with abbreviated arithmetic 'update' operators.
These operators are +=, -=, *=, and /=.
An interesting technique used here is to add two tokens where only one exists (e.g., += and then an extra + token). This helps later with the de-sugaring in the parser to immediately grab a token to the right binary operator to use.
This might somewhat complicate error messages (since some of these are 'phantom' tokens), though I've not come across any issues thus far. This approach seems to work stellarly.