Fix: prevented phantom SLASH token from being added if '/=' is not present.
Before this edit, a SLASH token is inserted wherever / is found (including at the beginning of comments.
This then leads to an error where a / seems to pop up randomly in the text to the Parser.
The current fix only adds this phantom token for the specific case of the compound assignment /= operator.