Change encapsulation approach for fields.
Prior to this commit, any fields declared in a constructor would be private, and thus public fields must be declared outside of the constructor.
Now, all fields (declared in the constructor or elsewhere) are public by default, and fields in the constructor can be made private with the 'safe' keyword.