Fix: formatting of saved input line(s) in runPrompt().
We cannot use originalLen (as before) if the new input happens to be empty, or we get an IndexError.
We also need to replace tabs with spaces manually in our REPL input (content and saved lines) for the error line-printing to be properly formatted.
Also unconditionally enabled line no., position, and error-line printing for files.